:refactor: db operations
