+++ backend/app/api/v1/routes/users.py
@@ -2,14 +2,14 @@ from fastapi import APIRouter, Request, Depends, Response, encoders
 import typing as t
 
 from db.session import get_db
-from db.crud import (
+from db.crud.users import (
     get_users,
     get_user,
     create_user,
     delete_user,
     edit_user,
 )
-from db.schemas import UserCreate, UserEdit, User, UserOut
+from db.schemas.users import UserCreate, UserEdit, User
 from core.auth import get_current_active_user, get_current_active_superuser
 
 users_router = r = APIRouter()
