+++ backend/app/db/schemas/projects.py
@@ -1,56 +1,6 @@
 from pydantic import BaseModel
 import typing as t
 
-
-### SCHEMAS FOR USERS ###
-
-
-class UserBase(BaseModel):
-    email: str
-    is_active: bool = True
-    is_superuser: bool = False
-    first_name: t.Optional[str] = None
-    last_name: t.Optional[str] = None
-
-
-class UserOut(UserBase):
-    pass
-
-
-class UserCreate(UserBase):
-    password: str
-
-    class Config:
-        orm_mode = True
-
-
-class UserEdit(UserBase):
-    password: t.Optional[str] = None
-
-    class Config:
-        orm_mode = True
-
-
-class User(UserBase):
-    id: int
-
-    class Config:
-        orm_mode = True
-
-
-### SCHEMAS FOR TOKENS ###
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    email: str = None
-    permissions: str = "user"
-
-
 ### SCHEMAS FOR PROJECTS ###
 
 
@@ -60,6 +10,7 @@ class Socials(BaseModel):
     discord: t.Optional[str]
     github: t.Optional[str]
     website: t.Optional[str]
+
 
 class CreateAndUpdateProject(BaseModel):
     name: str
