+++ backend/app/db/schemas/users.py
@@ -1,0 +1,38 @@
+from pydantic import BaseModel
+import typing as t
+
+
+### SCHEMAS FOR USERS ###
+
+
+class UserBase(BaseModel):
+    email: str
+    is_active: bool = True
+    is_superuser: bool = False
+    first_name: t.Optional[str] = None
+    last_name: t.Optional[str] = None
+
+
+class UserOut(UserBase):
+    pass
+
+
+class UserCreate(UserBase):
+    password: str
+
+    class Config:
+        orm_mode = True
+
+
+class UserEdit(UserBase):
+    password: t.Optional[str] = None
+
+    class Config:
+        orm_mode = True
+
+
+class User(UserBase):
+    id: int
+
+    class Config:
+        orm_mode = True
